feat(frontend): WorkflowEditor loads transform referenced in url
Dataset component wraps the WorkflowEditor component. The dataset component extracts a qriRef from router params, and provides it to the WorkflowEditor. The Dataset component also fires off an API call to load the dataset that matches the ref whenever it changes.

Down in Workflow editor we fire an event to update the ref the workflow section of the state tree is tuned to with an effect hook. The workflow reducer listens for API_DATASET_SUCCESS actions and if the ref of the api response matches the ref the workflow reducer is tuned to, it updates the workflow transform script.

There are two bugs not handled yet:
* if the user has edited the script, it would get clobbered if a matching API fetch action completed while they were mid edit. The right solution there is to have *two* transform scripts in the workflow reducer. one we "reset" to, and another for user edits. successful api responses should write to the "reset" slot.
* the references don't confirm the version loaded is HEAD. successfully loading a historical commit will lead to the wrong script loaded into the workflow reducer