refactor(`Workflow`): add `clearRun`, `setRunType` and `setRunButtonType` actions
In order to model multiple `run` type states, we need a `runType` (which models the currently displaying run type, and the `runButtonType` (which models if we are displaying "dry run" or "run and save" buttons). These get out of sync if the user "deploys" w/ a "save", which will get the `runType` and `runButtonType` out of sync.

When a user switches between the "dry run" and "run and save" options, or clicks on the "dry run" or "run and save" button, we need to clear the run (aka remove the `events` from the `workflow`), and ensure the run type matches the run button type.