fix(`NewDataset`): ensure resulting `Dataset` only has components that exist
each `NewCOMPONENT` function returns that corresponding component with a `qri` tag. Using each function without checking to see if the component exists means that even if the component didn't exist on the incoming dataset, that component will exist (with only the `qri` field) in the outgoing dataset.

Adjusted so that only components that exist in the incoming records get added to the outgoing dataset.