fix(`Body`): ensure columns and headers match
In `qri-desktop`, the codebase that the `ds_components` code is pulled from, when we fetch a `Body`, we pre-pend each row with an index. Rather that prepend the body, we should instead align the row number using the pagination information and the array index.

Until we have pagination/fetching, we rely only on the array index to give a row number.