test(`e2e`): fix open handlers error
Looks like the `--detectOpenHandles` flag was not actually detecting any open handles in our case, but instead hanging indefinitely. There was an open handle: our previous `frontend` `ChildProcess` was not being closed. However, a simple `frontend.kill()` was not working, because the actual command `yarn start` was actually spawned in its own process, one that cannot be controlled from the `frontend` process.

The solution (a little bit of a hack) is to have the `frontend` process spawned in `detached` mode, this forces the spawned process to not be attached to the main process, but instead go in a new group of processes. Any "grandchild" processes are then grouped under this same process. We can then kill all the processes using the `pid` of the spawned process.

To keep this noise outside of our e2e test, we now have a very simple `TestQrimaticFrontend` class, with a `close` method that handles killing the process for us.