refactor(*): add ids for e2e test selectors
First round of adding the "happy path" to the e2e tests.

Adjusted the relevant html elements to contain the correct ids, so that puppeteer can more easily find them.

Commented out the portions that are not currently possible in the app.

Currently struggling with interacting with monaco editor using puppeteer, when we get to the steps where we need to type into a monaco editor cell, the action does not happen. My suspicion is we need to interact by using focus and submitting keys, rather than using the usual puppeteer `page.type()` api.