refactor(`DatasetComponents`): takes a `selectedComponent` and `setSelectedComponent`
Decided the cleanest/fastest way to port the ds_components/DatasetComponents component was to isolate it from the logic of fetching the dataset or setting the selected component, for now.

We may decide that we want it to encapsulate all the logic needed to work on its own, (especially when we get pagination involved and then need to pass down a function that allows us to fetch new pages of the body, and/or how to do editing although I have some thoughts about that). However, isolating it this way means we open an easier path to using the Component Viewer in different parts of the app or porting it to other projects in the future.