refactor(`DatasetsTable`): each row's `hamburger` needs `row` information
Moving the creation of the `hamburgerItems` into the `cell` function of the hamburger cell. Since each hamburger item is going to need information from each row in order to construct `onClick` functions, moved this from a function component level variable down closer to the actual call to `<DropdownMenu>`

Also, made notes about where we are mocking `run` status information. This should really be co-located in one space as much as possible & outside of the component itself, but until the data model settles, it feels like this would get refactored a bunch anyway. So in the interest of time, using extensive `TODO`s as a way to remember what needs to be refactored/removed