feat(remove dataset): add `removeDataset` action to dataset state tree
Decided to add this to the `dataset` state tree, rather than the `collection` state since... well it's called `removeDataset`, and the `dataset` state tree is going to have to do more manipulation if a dataset that is being viewed is deleted.

The way we need to get at the underlying promise in the `ApiActionThunk` is a bit weird, and makes me think that something is not aligned in the most optimal way in our `api`.