feat(`Workflow`): add `LatestStart`, `LatestEnd`, and `Status` fields
`Workflow`s are not the same thing as `Run`s. A `Workflow` contains a `Run`, but it may have other things that a `Workflow` keeps track of (such as the status of the `OnComplete` actions, when they are implemented)that causes the `Start`, `End`, and `Status` to be different than those of the contained `Run`. As such, we need the `Workflow` to keep track of its own `LatestStart`, `LatestEnd`, and `Status`.