feat(`Store`): add `Subscribe` function that listens for Workflow events
When you create a new `Store`, the `Store` is now subscribed to the workflow events.

Currently we handle `ETWorkflowStarted` and `ETWorkflowCompleted` events, putting the updated workflows that are emitted with the event into the `Store`.

We expect the `RunWorkflow` event to adjust the emitted workflows with the most recent workflow information (such as `Status`, `LatestStart` or `LatestEnd`).

A `NewMemStore` and `NewFileStore` now require a `bus` in order to create a new `Store`