refactor(`Store`): un-export `MemStore` & `FileStore`
We need to ensure that a `memStore` & `fileStore` are guaranteed to be subscribed to the `Workflow` events. We do that by un-exporting `MemStore` and `FileStore` and forcing consumers outside of this package to use `NewMemStore` or `NewFileStore` if they need a `Store`.

This commit also removes `Subscribe` from the `Store` interface & exports and explains `EventHandlerFromStore` that subscribes a store to the necessary/expected event handlers.