feat(`Collection`): add `loadRunningCollection` & adjust `selectCollection`
We need to not only keep a record of the `WorkflowInfo`s that we fetch, we need to also keep an ordered list of  `ids` (workflow IDs) and `running` ids (the workflow ids of currently running datasets.

The ordered `running` and `ids` list helps us correctly order the `WorkflowInfo`s, while the collection of `WorkflowInfo` records allows us to update the `WorkflowInfo` if there are any changes to a `WorkflowInfo`

`running` workflow infos should always be returned before `ids` workflow infos. `running` workflows are ordered by `latestStart`, while `ids` are ordered by `commitTime`. Once a workflow has made it onto the `running` list, we should keep it there so it is placed where the user can see it, even if it has failed while running. We can change this behavior (or rather the natural behavior will be what we want to occur), once we are able to order datasets by `lastActivity`, rather than `CommitTime` (which will account for failures)