fix(`WorkflowStatus`): surface 'unchanged' as a `WorkflowStatus`
'unchanged' implies that the run was successful, but there were no changes to the underlying dataset.

I'm hesitant to surface this option, however. When we add `onComplete` events, I'm not sure how "no changes" and these options will interplay. There will be some interesting logic, depending on if a user has `onComplete` events that run even if a dataset has no changes. In which case the status of the workflow will be "succeeded", rather than "unchanged". We should just have these interactions well defined.