refactor(`sessionState`): use the current qri node's identity in the frontend
We need to do some hacks to ensure we don't keep enforcing bad behavior. Right now, we are using the idea that `me` implicitly translates to the identity of the underlying qri node. However, this causes bugs in a few places.

First, while `qri` understands what to do with the username "me", qrimatic does not know how to resolve "me". This is causing bugs where the workflow believes the dataset is called "me/dataset_name" rather than "underlying_username/dataset_name".

Second, our frontend has `me` hardcoded in some inconvenient places, such as in path and route names.

Even though we don't have a proper signup/login connection between frontend and backend yet, we can take one step closer by giving frontend a better sense of identity. This will also allow us to keep iterating on the UI of signing up/logging in & the flow of the app, while we don't yet have these proper connections in place.

This commit adds `fetchUsername` that calls to the `/me` endpoint & pulls the username from the response. `logIn` now requires that the username you attempt to log in with be the same username that the qri node identities as. We've also disabled the `signUp` function, which just auto-errors and lets the user know that signing up isn't currently working.