feat(frontend): component commit selector displays that commit
Re-work Dataset routes to support URLS to specific commits & fix a bug in
the process:

`react-router.useParams()` will only yeild the route params for the route the
current component is loaded within. Because of this we can't dispatch
`loadDataset` until we're in a component that has *all* route parameters,
including things like `/:fs` and `/:hash`.