feat(frontend): websocket connection healing
I ended up re-writing this four times, each with a slightly different apprach.

The first factoring had all websocket state in the app reducer. Having an "app"
reducer very similar to global state, and made it tempting to call the
connection a "backend connection" instead of a "websocket connection".

On the second factoring I moved all websocket-related code into the websocket
feature folder, and wrote a set of actions that used a "data" payload to
manipulate state in a reducer. All actions were of type "WS_CONNECTION_CHANGE",
and the change info was part of the data. I wasn't a fan of having the actions
themselves dictate business logic like the number of retries to attempt, etc.

Third factoring moved all of this logic into the reducer, and thinned the
actions to just a set of event constants without any payload ("WS_CONNECTED",
"WS_CONNECT_FAIL", etc), and the reducer owned how to update state according
to each action type. My plan was to have the websocket middleware fire these
actions, and keep state management in the reducer. But, the way redux middlware
integrates with reducers makes it difficult/impossible to fire actions & listen
for those actions within the same middlware (which, in hindsight, makes sense),
which meant the websocket middlware couldn't make use of state changes in the
websocket reducer.

So on the fourth go-round I ended up making the websocket middlware the owner
of websocket connection state, and used the approach from the second factoring
to store a copy of that state in the reducer. In this arrangement redux handles
a reducer that UI can listen to, and middlware doesn't have the additional
burden of state coordination with an external reducer.