feat(websocket): add authentication handshake to websocket connection
If we are already logged in, or have just successfully logged in, we need to authenticate our connection. This will let the backend know that it should send the events associated with the current session's profileID over the websocket connection.

If we ever logout, we need to unsubscribe to this connection.