feat(server): dataset preview routes add head & json-ld data to the html `head`
If a request comes to the server whose route matches `/:username/:name` (filtered for any potentially conflicting routes that we have in our app), we request the associated dataset preview from the backend and embed that data into the DOM.

The DOM data is structured in such a way that it can be found by google's dataset search. We've pulled much of that code from our `qri.cloud` repo.