refactor(collection): respond to events to update the collection
We want the collection to respond to currently running workflows. We can mirror the way that the backend collection responds to certain events in order to update the version infos in the current collection.

This commit add actions to the `collectionActions` that the `websocket` reducer can dispatch to alert the `collectionState` that certain activity is going on. These actions specifically respond to the `ETLogbookWriteCommit`, `ETLogbookWriteRun`, and `ETTransformStart` events.

The `collectionState` then uses the payloads passed by those particular events to update the collection.