feat(*): add `ExecutionMode` enum to switch on execution modes
Eventually, we want to be able to seamlessly use the frontend app locally in the browser, as a desktop application, or hosted on cloud (as we do now). Adding a `ExecutionMode` enum with options `LOCAL`, `DESKTOP`, and `CLOUD` & a global variable `APP_EXEC_MODE` that we can switch on to make mode specific adjustments.