refactor(store): use `API_BASE_CLOUD_API` on `identity` fetches
This is a hack to get the local implementation to behave like cloud, and to use the cloud api endpoints to fetch the correct data.

`API_BASE_CLOUD_API` will be used with any endpoints that include `identity/`

It can be set by `process.env.REACT_APP_BASE_CLOUD_API` and can be ignored when running in `CLOUD` execution mode.

The default endpoint is the development server api endpoints.